+++ src/main/java/nl/stil4m/mollie/domain/subpayments/paysafecard/CreatePaySafeCardPayment.java
@@ -1,15 +1,17 @@
 package nl.stil4m.mollie.domain.subpayments.paysafecard;
 
-import nl.stil4m.mollie.domain.subpayments.base.OptionedCreatePayment;
-
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
+import java.math.BigDecimal;
 import java.util.Map;
 import java.util.Optional;
 
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
+import nl.stil4m.mollie.domain.subpayments.base.OptionedCreatePayment;
+
 public class CreatePaySafeCardPayment extends OptionedCreatePayment<PaySafeCardOptions> {
 
-    public CreatePaySafeCardPayment(@Nonnull Double amount, @Nonnull String description, @Nonnull String redirectUrl, @Nonnull Optional<String> webhookUrl, @Nullable Map<String, Object> metadata, @Nonnull PaySafeCardOptions options) {
+    public CreatePaySafeCardPayment(@Nonnull BigDecimal amount, @Nonnull String description, @Nonnull String redirectUrl, @Nonnull Optional<String> webhookUrl, @Nullable Map<String, Object> metadata, @Nonnull PaySafeCardOptions options) {
         super("paysafecard", amount, description, redirectUrl, webhookUrl, metadata, options);
     }
 }
